refactor(dict): add generic html download function
